refactor: move ParseStateMachine into reader/internal
The parser package only held ParseStateMachine, which is internal and
used solely by SequenceParser. Co-locating it removes the single-file
package and improves cohesion.
